Move e2e tests into separate package
This requires refactoring of the library code, because e2e can't import
from main.
