+++ .gitignore
@@ -1,1 +1,1 @@
-etcdmon
+./etcdmon
