+++ Dockerfile
@@ -3,6 +3,7 @@ FROM golang:${GOVERSION}-alpine AS build
 
 WORKDIR /src
 COPY *.go go.mod go.sum ./
+COPY pkg/ ./pkg
 RUN --mount=type=cache,target=/go --mount=type=cache,target=/root/.cache/go-build \
 	go build -o /out/etcdmon .
