+++ pkg/etcdmon/k8s.go
@@ -1,4 +1,4 @@
-package main
+package etcdmon
 
 import (
 	"context"
